Refactoted to use redux
