+++ packages/pmml-editor/package.json
@@ -23,7 +23,10 @@
     "immer": "^7.0.8",
     "lodash": "^4.14.2",
     "@types/lodash": "^4.14.161",
-    "combine-reducer": "^1.0.2"
+    "combine-reducer": "^1.0.2",
+    "redux": "^4.0.5",
+    "react-redux": "^7.2.1",
+    "@types/react-redux": "^7.1.9"
   },
   "scripts": {
     "lint": "tslint -c ../../tslint.json 'src/**/*.{ts,tsx,js,jsx}'",
