+++ packages/pmml-editor/src/editor/MockHeaderUI.tsx
@@ -14,16 +14,19 @@
  * limitations under the License.
  */
 import * as React from "react";
-import { AppContext } from "./PMMLEditorContextProvider";
 import { Actions } from "./reducers/Actions";
 import { Timestamp, Title } from "./PMMLEditor";
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
+import { Header, PMML } from "@kogito-tooling/pmml-editor-marshaller";
 
 const style = {
   padding: "5px 5px 5px 5px"
 };
 
 const MockHeaderUI = () => {
-  const { state, dispatch } = React.useContext(AppContext);
+  const dispatch = useDispatch();
+  const typedUseSelector: TypedUseSelectorHook<PMML> = useSelector;
+  const header: Header = typedUseSelector(state => state.Header);
 
   const setHeaderDescription = (description: string) => {
     dispatch({
@@ -37,7 +40,7 @@ const MockHeaderUI = () => {
   return (
     <div style={style}>
       <Title title="Header" />
-      <input value={state.Header.description} onChange={e => setHeaderDescription(e.target.value)} placeholder="Name" />
+      <input value={header.description} onChange={e => setHeaderDescription(e.target.value)} placeholder="Name" />
       <Timestamp />
     </div>
   );
