+++ packages/pmml-editor/src/editor/MockVersionUI.tsx
@@ -14,16 +14,19 @@
  * limitations under the License.
  */
 import * as React from "react";
-import { AppContext } from "./PMMLEditorContextProvider";
 import { Actions } from "./reducers/Actions";
 import { Timestamp, Title } from "./PMMLEditor";
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
+import { PMML } from "@kogito-tooling/pmml-editor-marshaller";
 
 const style = {
   padding: "5px 5px 5px 5px"
 };
 
 const MockVersionUI = () => {
-  const { state, dispatch } = React.useContext(AppContext);
+  const dispatch = useDispatch();
+  const typedUseSelector: TypedUseSelectorHook<PMML> = useSelector;
+  const version: string = typedUseSelector(state => state.version);
 
   const setVersion = () => {
     dispatch({
@@ -40,7 +43,7 @@ const MockVersionUI = () => {
   return (
     <div style={style}>
       <Title title={"PMML"} />
-      <div>{state.version}</div>
+      <div>{version}</div>
       <button onClick={setVersion}>Set version</button>
       <Timestamp />
     </div>
